python_backend/requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
pandas==2.0.3
python-multipart==0.0.6
supabase==1.0.3
python-dotenv==1.0.0
openpyxl==3.1.2
